jap crud

+++ src/main/java/com/awsome/domain/Book.java
@@ -8,46 +8,46 @@ import java.io.Serializable;
 
 @Entity
 @Table(name = "t_book")
-public class Book implements Serializable{
-	@Id
-	@GeneratedValue
-	private Integer	id;
-	private String	name;
-	private String	num;
+public class Book implements Serializable {
+    @Id
+    @GeneratedValue
+    private Long id;
+    private String  name;
+    private String  num;
 
 
 
-	public Integer getId() {
-		return id;
-	}
+    public Long getId() {
+        return id;
+    }
 
 
 
-	public void setId(Integer id) {
-		this.id = id;
-	}
+    public void setId(Long id) {
+        this.id = id;
+    }
 
 
 
-	public String getName() {
-		return name;
-	}
+    public String getName() {
+        return name;
+    }
 
 
 
-	public void setName(String name) {
-		this.name = name;
-	}
+    public void setName(String name) {
+        this.name = name;
+    }
 
 
 
-	public String getNum() {
-		return num;
-	}
+    public String getNum() {
+        return num;
+    }
 
 
 
-	public void setNum(String num) {
-		this.num = num;
-	}
+    public void setNum(String num) {
+        this.num = num;
+    }
 }
